@@ -1,30 +1,30 @@
-﻿using System.Diagnostics;
+using System.Diagnostics;
 using System.Runtime.InteropServices;
 
 namespace SpotifyAPI.Web.Auth
 {
-    internal static class AuthUtil
+  internal static class AuthUtil
+  {
+    public static void OpenBrowser(string url)
     {
-        public static void OpenBrowser(string url)
-        {
 #if NETSTANDARD2_0
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                url = url.Replace("&", "^&");
-                Process.Start(new ProcessStartInfo("cmd", $"/c start {url}"));
-            }
-            else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-            {
-                Process.Start("xdg-open", url);
-            }
-            else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
-            {
-                Process.Start("open", url);
-            }
+      if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+      {
+        url = url.Replace("&", "^&");
+        Process.Start(new ProcessStartInfo("cmd", $"/c start {url}"));
+      }
+      else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+      {
+        Process.Start("xdg-open", url);
+      }
+      else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
+      {
+        Process.Start("open", url);
+      }
 #else
-            url = url.Replace("&", "^&");
-            Process.Start(new ProcessStartInfo("cmd", $"/c start {url}"));
+      url = url.Replace("&", "^&");
+      Process.Start(new ProcessStartInfo("cmd", $"/c start {url}"));
 #endif
-        }
     }
+  }
 }
