@@ -2,10 +2,10 @@ using System;
 
 namespace SpotifyAPI.Web.Examples.ASP.Models
 {
-    public class ErrorViewModel
-    {
-        public string RequestId { get; set; }
+  public class ErrorViewModel
+  {
+    public string RequestId { get; set; }
 
-        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-    }
+    public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+  }
 }
