@@ -3,8 +3,8 @@ using SpotifyAPI.Web.Models;
 
 namespace SpotifyAPI.Web.Examples.ASP.Models
 {
-    public class IndexModel
-    {
-        public Paging<SavedTrack> SavedTracks;
-    }
+  public class IndexModel
+  {
+    public Paging<SavedTrack> SavedTracks;
+  }
 }
