@@ -1,23 +1,23 @@
-﻿using System;
+using System;
 
 namespace SpotifyAPI.Web.Enums
 {
-    [Flags]
-    public enum AlbumType
-    {
-        [String("album")]
-        Album = 1,
+  [Flags]
+  public enum AlbumType
+  {
+    [String("album")]
+    Album = 1,
 
-        [String("single")]
-        Single = 2,
+    [String("single")]
+    Single = 2,
 
-        [String("compilation")]
-        Compilation = 4,
+    [String("compilation")]
+    Compilation = 4,
 
-        [String("appears_on")]
-        AppearsOn = 8,
+    [String("appears_on")]
+    AppearsOn = 8,
 
-        [String("album,single,compilation,appears_on")]
-        All = 16
-    }
+    [String("album,single,compilation,appears_on")]
+    All = 16
+  }
 }
