@@ -1,14 +1,14 @@
-﻿using System;
+using System;
 
 namespace SpotifyAPI.Web.Enums
 {
-    [Flags]
-    public enum FollowType
-    {
-        [String("artist")]
-        Artist = 1,
+  [Flags]
+  public enum FollowType
+  {
+    [String("artist")]
+    Artist = 1,
 
-        [String("user")]
-        User = 2
-    }
+    [String("user")]
+    User = 2
+  }
 }
