@@ -1,17 +1,17 @@
-﻿using System;
+using System;
 
 namespace SpotifyAPI.Web.Enums
 {
-    [Flags]
-    public enum RepeatState
-    {
-        [String("track")]
-        Track = 1,
+  [Flags]
+  public enum RepeatState
+  {
+    [String("track")]
+    Track = 1,
 
-        [String("context")]
-        Context = 2,
+    [String("context")]
+    Context = 2,
 
-        [String("off")]
-        Off = 4
-    }
+    [String("off")]
+    Off = 4
+  }
 }
