@@ -1,9 +1,9 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class ListResponse<T> : BasicModel
-    {
-        public List<T> List { get; set; }
-    }
+  public class ListResponse<T> : BasicModel
+  {
+    public List<T> List { get; set; }
+  }
 }
