@@ -1,11 +1,11 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class AvailabeDevices : BasicModel
-    {
-        [JsonProperty("devices")]
-        public List<Device> Devices { get; set; }
-    }
+  public class AvailabeDevices : BasicModel
+  {
+    [JsonProperty("devices")]
+    public List<Device> Devices { get; set; }
+  }
 }
