@@ -1,20 +1,20 @@
-﻿using Newtonsoft.Json;
 using System.Collections.Generic;
+using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class Category : BasicModel
-    {
-        [JsonProperty("href")]
-        public string Href { get; set; }
+  public class Category : BasicModel
+  {
+    [JsonProperty("href")]
+    public string Href { get; set; }
 
-        [JsonProperty("icons")]
-        public List<Image> Icons { get; set; }
+    [JsonProperty("icons")]
+    public List<Image> Icons { get; set; }
 
-        [JsonProperty("id")]
-        public string Id { get; set; }
+    [JsonProperty("id")]
+    public string Id { get; set; }
 
-        [JsonProperty("name")]
-        public string Name { get; set; }
-    }
+    [JsonProperty("name")]
+    public string Name { get; set; }
+  }
 }
