@@ -1,10 +1,10 @@
-﻿using Newtonsoft.Json;
+using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class CategoryList : BasicModel
-    {
-        [JsonProperty("categories")]
-        public Paging<Category> Categories { get; set; }
-    }
+  public class CategoryList : BasicModel
+  {
+    [JsonProperty("categories")]
+    public Paging<Category> Categories { get; set; }
+  }
 }
