@@ -1,10 +1,10 @@
-﻿using Newtonsoft.Json;
+using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class CategoryPlaylist : BasicModel
-    {
-        [JsonProperty("playlists")]
-        public Paging<SimplePlaylist> Playlists { get; set; }
-    }
+  public class CategoryPlaylist : BasicModel
+  {
+    [JsonProperty("playlists")]
+    public Paging<SimplePlaylist> Playlists { get; set; }
+  }
 }
