@@ -1,13 +1,13 @@
-﻿using Newtonsoft.Json;
+using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class FeaturedPlaylists : BasicModel
-    {
-        [JsonProperty("message")]
-        public string Message { get; set; }
+  public class FeaturedPlaylists : BasicModel
+  {
+    [JsonProperty("message")]
+    public string Message { get; set; }
 
-        [JsonProperty("playlists")]
-        public Paging<SimplePlaylist> Playlists { get; set; }
-    }
+    [JsonProperty("playlists")]
+    public Paging<SimplePlaylist> Playlists { get; set; }
+  }
 }
