@@ -1,10 +1,10 @@
-﻿using Newtonsoft.Json;
+using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class FollowedArtists : BasicModel
-    {
-        [JsonProperty("artists")]
-        public CursorPaging<FullArtist> Artists { get; set; }
-    }
+  public class FollowedArtists : BasicModel
+  {
+    [JsonProperty("artists")]
+    public CursorPaging<FullArtist> Artists { get; set; }
+  }
 }
