@@ -1,68 +1,68 @@
-﻿using Newtonsoft.Json;
 using System.Collections.Generic;
+using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class FullAlbum : BasicModel
-    {
-        [JsonProperty("album_type")]
-        public string AlbumType { get; set; }
+  public class FullAlbum : BasicModel
+  {
+    [JsonProperty("album_type")]
+    public string AlbumType { get; set; }
 
-        [JsonProperty("artists")]
-        public List<SimpleArtist> Artists { get; set; }
+    [JsonProperty("artists")]
+    public List<SimpleArtist> Artists { get; set; }
 
-        [JsonProperty("available_markets")]
-        public List<string> AvailableMarkets { get; set; }
+    [JsonProperty("available_markets")]
+    public List<string> AvailableMarkets { get; set; }
 
-        [JsonProperty("copyrights")]
-        public List<Copyright> Copyrights { get; set; }
+    [JsonProperty("copyrights")]
+    public List<Copyright> Copyrights { get; set; }
 
-        [JsonProperty("external_ids")]
-        public Dictionary<string, string> ExternalIds { get; set; }
+    [JsonProperty("external_ids")]
+    public Dictionary<string, string> ExternalIds { get; set; }
 
-        [JsonProperty("external_urls")]
-        public Dictionary<string, string> ExternalUrls { get; set; }
+    [JsonProperty("external_urls")]
+    public Dictionary<string, string> ExternalUrls { get; set; }
 
-        [JsonProperty("genres")]
-        public List<string> Genres { get; set; }
+    [JsonProperty("genres")]
+    public List<string> Genres { get; set; }
 
-        [JsonProperty("href")]
-        public string Href { get; set; }
+    [JsonProperty("href")]
+    public string Href { get; set; }
 
-        [JsonProperty("id")]
-        public string Id { get; set; }
+    [JsonProperty("id")]
+    public string Id { get; set; }
 
-        [JsonProperty("images")]
-        public List<Image> Images { get; set; }
-        
-        [JsonProperty("label")]
-        public string Label { get; set; }
+    [JsonProperty("images")]
+    public List<Image> Images { get; set; }
 
-        [JsonProperty("name")]
-        public string Name { get; set; }
+    [JsonProperty("label")]
+    public string Label { get; set; }
 
-        [JsonProperty("popularity")]
-        public int Popularity { get; set; }
+    [JsonProperty("name")]
+    public string Name { get; set; }
 
-        [JsonProperty("release_date")]
-        public string ReleaseDate { get; set; }
+    [JsonProperty("popularity")]
+    public int Popularity { get; set; }
 
-        [JsonProperty("release_date_precision")]
-        public string ReleaseDatePrecision { get; set; }
+    [JsonProperty("release_date")]
+    public string ReleaseDate { get; set; }
 
-        [JsonProperty("tracks")]
-        public Paging<SimpleTrack> Tracks { get; set; }
+    [JsonProperty("release_date_precision")]
+    public string ReleaseDatePrecision { get; set; }
 
-        [JsonProperty("restrictions")]
-        public Dictionary<string, string> Restrictions { get; set; }
+    [JsonProperty("tracks")]
+    public Paging<SimpleTrack> Tracks { get; set; }
 
-        [JsonProperty("total_tracks")]
-        public int TotalTracks { get; set; }
+    [JsonProperty("restrictions")]
+    public Dictionary<string, string> Restrictions { get; set; }
 
-        [JsonProperty("type")]
-        public string Type { get; set; }
+    [JsonProperty("total_tracks")]
+    public int TotalTracks { get; set; }
 
-        [JsonProperty("uri")]
-        public string Uri { get; set; }
-    }
+    [JsonProperty("type")]
+    public string Type { get; set; }
+
+    [JsonProperty("uri")]
+    public string Uri { get; set; }
+  }
 }
