@@ -1,71 +1,71 @@
-﻿using Newtonsoft.Json;
 using System.Collections.Generic;
+using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class FullTrack : BasicModel
-    {
-        [JsonProperty("album")]
-        public SimpleAlbum Album { get; set; }
+  public class FullTrack : BasicModel
+  {
+    [JsonProperty("album")]
+    public SimpleAlbum Album { get; set; }
 
-        [JsonProperty("artists")]
-        public List<SimpleArtist> Artists { get; set; }
+    [JsonProperty("artists")]
+    public List<SimpleArtist> Artists { get; set; }
 
-        [JsonProperty("available_markets")]
-        public List<string> AvailableMarkets { get; set; }
+    [JsonProperty("available_markets")]
+    public List<string> AvailableMarkets { get; set; }
 
-        [JsonProperty("disc_number")]
-        public int DiscNumber { get; set; }
+    [JsonProperty("disc_number")]
+    public int DiscNumber { get; set; }
 
-        [JsonProperty("duration_ms")]
-        public int DurationMs { get; set; }
+    [JsonProperty("duration_ms")]
+    public int DurationMs { get; set; }
 
-        [JsonProperty("explicit")]
-        public bool Explicit { get; set; }
+    [JsonProperty("explicit")]
+    public bool Explicit { get; set; }
 
-        [JsonProperty("external_ids")]
-        public Dictionary<string, string> ExternalIds { get; set; }
+    [JsonProperty("external_ids")]
+    public Dictionary<string, string> ExternalIds { get; set; }
 
-        [JsonProperty("external_urls")]
-        public Dictionary<string, string> ExternUrls { get; set; }
+    [JsonProperty("external_urls")]
+    public Dictionary<string, string> ExternUrls { get; set; }
 
-        [JsonProperty("href")]
-        public string Href { get; set; }
+    [JsonProperty("href")]
+    public string Href { get; set; }
 
-        [JsonProperty("id")]
-        public string Id { get; set; }
+    [JsonProperty("id")]
+    public string Id { get; set; }
 
-        [JsonProperty("name")]
-        public string Name { get; set; }
+    [JsonProperty("name")]
+    public string Name { get; set; }
 
-        [JsonProperty("popularity")]
-        public int Popularity { get; set; }
+    [JsonProperty("popularity")]
+    public int Popularity { get; set; }
 
-        [JsonProperty("preview_url")]
-        public string PreviewUrl { get; set; }
+    [JsonProperty("preview_url")]
+    public string PreviewUrl { get; set; }
 
-        [JsonProperty("track_number")]
-        public int TrackNumber { get; set; }
+    [JsonProperty("track_number")]
+    public int TrackNumber { get; set; }
 
-        [JsonProperty("restrictions")]
-        public Dictionary<string, string> Restrictions { get; set; }
+    [JsonProperty("restrictions")]
+    public Dictionary<string, string> Restrictions { get; set; }
 
-        [JsonProperty("type")]
-        public string Type { get; set; }
+    [JsonProperty("type")]
+    public string Type { get; set; }
 
-        [JsonProperty("uri")]
-        public string Uri { get; set; }
+    [JsonProperty("uri")]
+    public string Uri { get; set; }
 
-        /// <summary>
-        ///     Only filled when the "market"-parameter was supplied!
-        /// </summary>
-        [JsonProperty("is_playable")]
-        public bool? IsPlayable { get; set; }
+    /// <summary>
+    ///     Only filled when the "market"-parameter was supplied!
+    /// </summary>
+    [JsonProperty("is_playable")]
+    public bool? IsPlayable { get; set; }
 
-        /// <summary>
-        ///     Only filled when the "market"-parameter was supplied!
-        /// </summary>
-        [JsonProperty("linked_from")]
-        public LinkedFrom LinkedFrom { get; set; }
-    }
+    /// <summary>
+    ///     Only filled when the "market"-parameter was supplied!
+    /// </summary>
+    [JsonProperty("linked_from")]
+    public LinkedFrom LinkedFrom { get; set; }
+  }
 }
