@@ -1,10 +1,10 @@
-﻿using Newtonsoft.Json;
+using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class NewAlbumReleases : BasicModel
-    {
-        [JsonProperty("albums")]
-        public Paging<SimpleAlbum> Albums { get; set; }
-    }
+  public class NewAlbumReleases : BasicModel
+  {
+    [JsonProperty("albums")]
+    public Paging<SimpleAlbum> Albums { get; set; }
+  }
 }
