@@ -1,11 +1,11 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class RecommendationSeedGenres : BasicModel
-    {
-         [JsonProperty("genres")]
-         public List<string> Genres { get; set; }
-    }
+  public class RecommendationSeedGenres : BasicModel
+  {
+    [JsonProperty("genres")]
+    public List<string> Genres { get; set; }
+  }
 }
