@@ -1,14 +1,14 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class Recommendations : BasicModel
-    {
-        [JsonProperty("seeds")]
-        public List<RecommendationSeed> Seeds { get; set; }
+  public class Recommendations : BasicModel
+  {
+    [JsonProperty("seeds")]
+    public List<RecommendationSeed> Seeds { get; set; }
 
-        [JsonProperty("tracks")]
-        public List<SimpleTrack> Tracks { get; set; }
-    }
+    [JsonProperty("tracks")]
+    public List<SimpleTrack> Tracks { get; set; }
+  }
 }
