@@ -1,11 +1,11 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class SeveralAlbums : BasicModel
-    {
-        [JsonProperty("albums")]
-        public List<FullAlbum> Albums { get; set; }
-    }
+  public class SeveralAlbums : BasicModel
+  {
+    [JsonProperty("albums")]
+    public List<FullAlbum> Albums { get; set; }
+  }
 }
