@@ -3,9 +3,9 @@ using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class SeveralArtists : BasicModel
-    {
-        [JsonProperty("artists")]
-        public List<FullArtist> Artists { get; set; }
-    }
+  public class SeveralArtists : BasicModel
+  {
+    [JsonProperty("artists")]
+    public List<FullArtist> Artists { get; set; }
+  }
 }
