@@ -1,11 +1,11 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class SeveralTracks : BasicModel
-    {
-        [JsonProperty("tracks")]
-        public List<FullTrack> Tracks { get; set; }
-    }
+  public class SeveralTracks : BasicModel
+  {
+    [JsonProperty("tracks")]
+    public List<FullTrack> Tracks { get; set; }
+  }
 }
