@@ -1,53 +1,53 @@
-﻿using Newtonsoft.Json;
 using System.Collections.Generic;
+using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class SimpleAlbum : BasicModel
-    {
-        [JsonProperty("album_group")]
-        public string AlbumGroup { get; set; }
+  public class SimpleAlbum : BasicModel
+  {
+    [JsonProperty("album_group")]
+    public string AlbumGroup { get; set; }
 
-        [JsonProperty("album_type")]
-        public string AlbumType { get; set; }
+    [JsonProperty("album_type")]
+    public string AlbumType { get; set; }
 
-        [JsonProperty("artists")]
-        public List<SimpleArtist> Artists { get; set; }
+    [JsonProperty("artists")]
+    public List<SimpleArtist> Artists { get; set; }
 
-        [JsonProperty("available_markets")]
-        public List<string> AvailableMarkets { get; set; }
+    [JsonProperty("available_markets")]
+    public List<string> AvailableMarkets { get; set; }
 
-        [JsonProperty("external_urls")]
-        public Dictionary<string, string> ExternalUrls { get; set; }
+    [JsonProperty("external_urls")]
+    public Dictionary<string, string> ExternalUrls { get; set; }
 
-        [JsonProperty("href")]
-        public string Href { get; set; }
+    [JsonProperty("href")]
+    public string Href { get; set; }
 
-        [JsonProperty("id")]
-        public string Id { get; set; }
+    [JsonProperty("id")]
+    public string Id { get; set; }
 
-        [JsonProperty("images")]
-        public List<Image> Images { get; set; }
+    [JsonProperty("images")]
+    public List<Image> Images { get; set; }
 
-        [JsonProperty("name")]
-        public string Name { get; set; }
+    [JsonProperty("name")]
+    public string Name { get; set; }
 
-        [JsonProperty("release_date")]
-        public string ReleaseDate { get; set; }
+    [JsonProperty("release_date")]
+    public string ReleaseDate { get; set; }
 
-        [JsonProperty("release_date_precision")]
-        public string ReleaseDatePrecision { get; set; }
+    [JsonProperty("release_date_precision")]
+    public string ReleaseDatePrecision { get; set; }
 
-        [JsonProperty("restrictions")]
-        public Dictionary<string, string> Restrictions { get; set; }
+    [JsonProperty("restrictions")]
+    public Dictionary<string, string> Restrictions { get; set; }
 
-        [JsonProperty("total_tracks")]
-        public int TotalTracks { get; set; }
+    [JsonProperty("total_tracks")]
+    public int TotalTracks { get; set; }
 
-        [JsonProperty("type")]
-        public string Type { get; set; }
+    [JsonProperty("type")]
+    public string Type { get; set; }
 
-        [JsonProperty("uri")]
-        public string Uri { get; set; }
-    }
+    [JsonProperty("uri")]
+    public string Uri { get; set; }
+  }
 }
