@@ -1,10 +1,10 @@
-﻿using Newtonsoft.Json;
+using Newtonsoft.Json;
 
 namespace SpotifyAPI.Web.Models
 {
-    public class Snapshot : BasicModel
-    {
-        [JsonProperty("snapshot_id")]
-        public string SnapshotId { get; set; }
-    }
+  public class Snapshot : BasicModel
+  {
+    [JsonProperty("snapshot_id")]
+    public string SnapshotId { get; set; }
+  }
 }
